Add Azure Functions infrastructure and ArmstrongNumbersFunction
This commit adds all necessary files to support Azure Functions in the application. It also introduces the ArmstrongNumbersFunction, an HTTP trigger function that processes a request and outputs whether the provided numerical input is an Armstrong number or not. New configurations and settings relevant to Azure Functions and the ArmstrongNumbersFunction were also included in the project.