Implement Pig Latin translation and update other functions
A new Pig Latin translation functionality was added to the solution. This <PigLatin> function takes a sentence as input and translates it to Pig Latin. Modifications were also made to the serialization in <PangramFunction>, conditions in <ScrabbleScoreFunction>, and logger usage in <ChangeFunction>. A new HTTP test for the functions was implemented as well.